requirements.txt
pandas
scikit-learn
joblib
tk
requests
beautifulsoup4
newspaper3k
lxml_html_clean
nltk
